+++ packages/client/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.15",
     "@mui/material": "^5.15.14",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.1.1",
