+++ packages/client/src/App.js
@@ -5,13 +5,12 @@ import Dashboard from './containers/Dashboard/Dashboard';
 // import { LandingPageContainer } from './containers/LandingPage/LandingPage';
 import { PageNotFound } from './containers/PageNotFound/PageNotFound.Container';
 
-
 function App() {
   return (
     <div className="app">
       <Router>
         <Routes>
-         {/* <Route path="/" element={<LandingPageContainer />} /> */}
+          {/* <Route path="/" element={<LandingPageContainer />} /> */}
           <Route path="/dashboard" element={<Dashboard />} />
           <Route path="*" element={<PageNotFound />} />
         </Routes>
