+++ packages/client/src/containers/Dashboard/Dashboard.js
@@ -4,10 +4,11 @@ import NavigationBar from '../../components/Dashboard/NavigationBar';
 import logo from './logo-dark.svg';
 import placeholder from './placeholder.png';
 import './Dashboard.css';
+import { TeamMembers } from '../../components/Dashboard/TeamMembers';
 
 const Dashboard = () => {
   return (
-    <div className= "dashboard-container">
+    <div className="dashboard-container">
       <div className="left-nav-bar-container">
         <img className="logo" src={logo} alt="Logo" />
         <NavigationBar />
@@ -16,20 +17,31 @@ const Dashboard = () => {
         <div className="top-nav-bar">
           <div className="heading-text">
             <Typography variant="h4">Dashboard</Typography>
-            <Typography className= "sub-heading" variant="body1">Welcome back, Meraj</Typography>
+            <Typography className="sub-heading" variant="body1">
+              Welcome back, Meraj
+            </Typography>
           </div>
 
           <div className="image-container">
-            <img className="placeholder-image" src={placeholder} alt="Placeholder" />
+            <img
+              className="placeholder-image"
+              src={placeholder}
+              alt="Placeholder"
+            />
             <div className="name-mail-container">
               <Typography variant="h6">Meraj</Typography>
-              <Typography className= "subtext" variant="body2">meraj@gmail.com</Typography>
+              <Typography className="subtext" variant="body2">
+                meraj@gmail.com
+              </Typography>
             </div>
           </div>
         </div>
+        <div>
+          <TeamMembers />
+        </div>
       </div>
-      </div>
+    </div>
   );
-}
+};
 
 export default Dashboard;
