+++ packages/client/src/containers/LandingPage/AddTeamMemberModal.css
@@ -1,0 +1,55 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  padding: 30px;
+  background-color: #fff;
+  width: 100%;
+  margin-top: 15px;
+}
+
+.add-member {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-around;
+  width: 100%;
+}
+
+.add-member input {
+  flex: 1;
+  margin-right: 10px;
+  padding: 8px;
+  border: 1px solid #ccc;
+  border-radius: 3px;
+}
+
+.add-member .add-button,
+.add-member .cancel-button {
+  padding: 8px 16px;
+  border: none;
+  border-radius: 3px;
+  cursor: pointer;
+}
+
+.add-member .add-button {
+  background-color: #007bff;
+  color: white;
+}
+
+.add-member .cancel-button {
+  background-color: #f0f0f0;
+  color: #333;
+}
+
+.add-member .cancel-button:hover {
+  background-color: #ddd;
+}
+
+.add-member button:last-child {
+  margin-left: 10px;
+}
+
+.add-member button:hover {
+  background-color: #0056b3;
+}
