--- packages/client/src/containers/LandingPage/Dashboard.css
@@ -1,5 +1,0 @@
-ul {
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-}
